backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
python-jose[cryptography]
pydantic-settings
pydantic[email]
redis
alembic
prometheus-fastapi-instrumentator
pytest
pytest-asyncio
httpx
langchain
langchain-text-splitters
langchain-google-genai
langchain-community
faiss-cpu
pypdf
google-generativeai
python-dotenv
python-multipart
